Moving GithubBranch to use yaml config. Working on #78
