+++ spec/DevBoard/Github/Branch/Entity/GithubBranchRepositorySpec.php
@@ -1,5 +1,5 @@
 <?php
-namespace spec\DevBoard\Github\Branch;
+namespace spec\DevBoard\Github\Branch\Entity;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
@@ -10,7 +10,7 @@ class GithubBranchRepositorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('DevBoard\Github\Branch\GithubBranchRepository');
+        $this->shouldHaveType('DevBoard\Github\Branch\Entity\GithubBranchRepository');
     }
 
     public function let(EntityManager $em, ClassMetadata $classMetadata)
