+++ src/DevBoard/Behat/Github/Branch/FixtureContext.php
@@ -3,7 +3,7 @@ namespace DevBoard\Behat\Github\Branch;
 
 use Behat\Gherkin\Node\TableNode;
 use DevBoard\Behat\Github\Repo\DataTrait as RepoDataTrait;
-use DevBoard\Github\Branch\GithubBranch;
+use DevBoard\Github\Branch\Entity\GithubBranch;
 use Resources\Behat\DefaultContext;
 
 /**
